Gun Settings component start - Components reorganised - New settings and insulin component - Settings component gun.set working gun.get not working
