+++ InitialDesign/components/gunComponent.js
@@ -1,8 +1,7 @@
 import React from "react";
 import { StyleSheet, View, Text, TextInput, Button } from "react-native";
-import Gun from 'gun/gun.js' // or use the minified version 'gun/gun.min.js'
-const gun = new Gun('https://diabetesappfontysgroep3.herokuapp.com/gun') // or use your own GUN server
-//Component.prototype.$gun = gun
+import Gun from 'gun/gun.js' 
+const gun = new Gun('https://diabetesappfontysgroep3.herokuapp.com/gun') 
 
 export default class CustomGun extends React.Component {
